requirements.txt
langchain>=0.2.11
langgraph>=0.2.34
langchain-openai>=0.1.20
langchain-community>=0.2.10
langchain-text-splitters>=0.2.2
faiss-cpu>=1.8.0.post1
pydantic>=2.7.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
